feat: implement discussionAllInterview model method for retrieving interview details
